[Refactor#70] 인증 관련 파일 home 모듈에서 approval 모듈로 이동
